Base types on generics, not global JSX namespace
For a while now, JSX frameworks don’t have to declare the global `JSX`
namespace anymore. This is a significant improvement when using multiple
JSX frameworks in the same project. React has deprecated the global
`JSX` namespace. Preact doesn’t even define it anymore.

This change replaces the use of the `JSX` namespace with generics, to
make it compatible with modern JSX frameworks.