Fix type error if `JSX` is not defined
React 19 no longer defined `JSX`.
It prefers users to use `React.JSX` instead.
Without it,
it becomes impossible to infer what the types of props for particular
components are,
what the know tag names are,
what the “result” of `jsx()` is.

This PR is a patch to prevent errors when `JSX` is not there
while still *allowing* `JSX` to exist and for this package to be typed nicely.

Related-to GH-4.
Closes GH-6.